requirements.txt
google-genai>=1.0.0
pdfplumber>=0.11.0
pymupdf>=1.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
streamlit>=1.40.0
plotly>=5.24.0
pandas>=2.0.0
pytest>=8.0.0
